+++ lib/segment/benches/numeric_index_check_values.rs
@@ -6,7 +6,7 @@ use rand::prelude::StdRng;
 use rand::{RngExt, SeedableRng};
 use segment::common::operation_error::OperationResult;
 use segment::index::field_index::numeric_index::NumericIndexRead;
-use segment::index::field_index::numeric_index::mmap_numeric_index::UniversalNumericIndex;
+use segment::index::field_index::numeric_index::universal_numeric_index::UniversalNumericIndex;
 use segment::index::field_index::numeric_index::mutable_numeric_index::InMemoryNumericIndex;
 use tempfile::Builder;
