+++ lib/segment/src/index/field_index/numeric_index/immutable_numeric_index/lifecycle.rs
@@ -5,7 +5,7 @@ use common::types::PointOffsetType;
 use gridstore::Blob;
 
 use super::super::Encodable;
-use super::super::mmap_numeric_index::UniversalNumericIndex;
+use super::super::universal_numeric_index::UniversalNumericIndex;
 use super::super::mutable_numeric_index::InMemoryNumericIndex;
 use super::{ImmutableNumericIndex, NumericKeySortedVec};
 use crate::common::Flusher;
