+++ lib/segment/src/index/field_index/numeric_index/mod.rs
@@ -2,7 +2,7 @@ mod builders;
 mod encodable;
 pub mod immutable_numeric_index;
 mod lifecycle;
-pub mod mmap_numeric_index;
+pub mod universal_numeric_index;
 pub mod mutable_numeric_index;
 mod numeric_field_index;
 pub mod numeric_index_read;
