+++ lib/segment/src/index/field_index/numeric_index/mutable_numeric_index/lifecycle.rs
@@ -9,7 +9,7 @@ use gridstore::{Blob, Gridstore};
 
 use super::super::Encodable;
 use super::super::lifecycle::{HISTOGRAM_MAX_BUCKET_SIZE, HISTOGRAM_PRECISION};
-use super::super::mmap_numeric_index::UniversalNumericIndex;
+use super::super::universal_numeric_index::UniversalNumericIndex;
 use super::super::numeric_index_read::NumericIndexRead;
 use super::{InMemoryNumericIndex, MutableNumericIndex, default_gridstore_options};
 use crate::common::Flusher;
