+++ lib/segment/src/index/field_index/tests/histogram_tests.rs
@@ -2,6 +2,7 @@ use std::cell::Cell;
 use std::collections::BTreeSet;
 use std::collections::Bound::{Excluded, Included, Unbounded};
 
+use common::universal_io::MmapFs;
 use itertools::Itertools;
 use rand::prelude::StdRng;
 use rand::{RngExt, SeedableRng};
@@ -278,6 +279,6 @@ fn test_save_load_histogram() {
         .unwrap();
     histogram.save(dir.path()).unwrap();
 
-    let loaded_histogram = Histogram::<f64>::load(dir.path()).unwrap();
+    let loaded_histogram = Histogram::<f64>::load_via(&MmapFs, dir.path()).unwrap();
     assert_eq!(histogram, loaded_histogram);
 }
